Update install example to avoid directory assumptions
The previous instruction examples assumed that the directory `~/github` existed, and did not include any check to create it if the directory did not exist. I updated to install in whatever directory the user is current in. 

I also updated to make it clear how the CUDA version is specified.